fix(thumbor.ts): rename thubmor -> thumbor
- typo in filename
